+++ src/window_main/components/collection/filters.tsx
@@ -4,7 +4,7 @@ import React from "react";
 import { ColumnInstance, Row } from "react-table";
 import { CARD_RARITIES, MANA } from "../../../shared/constants";
 import db from "../../../shared/database";
-import { RaritySymbol, SetSymbol, TypeSymbol } from "../display";
+import { RaritySymbol, SetSymbol, TypeSymbol } from "../misc/display";
 import { BinaryColumnFilter, BinaryFilterValue } from "../tables/filters";
 import { useMultiSelectFilter } from "../tables/hooks";
 import { MultiSelectFilterProps } from "../tables/types";
