+++ src/window_main/components/events/EventsTable.tsx
@@ -5,8 +5,8 @@ import { EVENTS_TABLE_MODE } from "../../../shared/constants";
 import pd from "../../../shared/PlayerData";
 import Aggregator, { AggregatorFilters } from "../../aggregator";
 import { ListItemEvent } from "../list-item/ListItemEvent";
-import MatchResultsStatsPanel from "../MatchResultsStatsPanel";
-import ResizableDragger from "../ResizableDragger";
+import MatchResultsStatsPanel from "../misc/MatchResultsStatsPanel";
+import ResizableDragger from "../misc/ResizableDragger";
 import {
   ArchivedCell,
   ArchiveHeader,
