+++ src/window_main/components/matches/MatchesTable.tsx
@@ -4,8 +4,8 @@ import { DATE_SEASON, MATCHES_TABLE_MODE } from "../../../shared/constants";
 import pd from "../../../shared/PlayerData";
 import Aggregator, { AggregatorFilters } from "../../aggregator";
 import ListItemMatch from "../list-item/ListItemMatch";
-import MatchResultsStatsPanel from "../MatchResultsStatsPanel";
-import ResizableDragger from "../ResizableDragger";
+import MatchResultsStatsPanel from "../misc/MatchResultsStatsPanel";
+import ResizableDragger from "../misc/ResizableDragger";
 import {
   ArchivedCell,
   ArchiveHeader,
