+++ src/window_main/components/misc/CardList.tsx
@@ -1,10 +1,10 @@
 /* eslint-disable react/prop-types */
 import _ from "lodash";
 import React from "react";
-import db from "../../shared/database";
-import CardTile from "../../shared/CardTile";
-import CardsList from "../../shared/cardsList";
-import { CardObject } from "../../types/Deck";
+import db from "../../../shared/database";
+import CardTile from "../../../shared/CardTile";
+import CardsList from "../../../shared/cardsList";
+import { CardObject } from "../../../types/Deck";
 
 interface CardListProps {
   list: CardsList;
