+++ src/window_main/components/misc/Slider.tsx
