+++ src/window_main/components/misc/display.tsx
@@ -1,12 +1,12 @@
 import React from "react";
 import styled from "styled-components";
-import { MANA } from "../../shared/constants";
-import db from "../../shared/database";
-import { get_rank_index_16 as getRankIndex16 } from "../../shared/util";
-import { getTagColor } from "../renderer-util";
-import AutosuggestInput from "./tables/AutosuggestInput";
-import { TagCounts } from "./tables/types";
-import useColorPicker from "../hooks/useColorPicker";
+import { MANA } from "../../../shared/constants";
+import db from "../../../shared/database";
+import { get_rank_index_16 as getRankIndex16 } from "../../../shared/util";
+import { getTagColor } from "../../renderer-util";
+import AutosuggestInput from "../tables/AutosuggestInput";
+import { TagCounts } from "../tables/types";
+import useColorPicker from "../../hooks/useColorPicker";
 
 export const ArtTileHeader = styled.div`
   width: 200px;
