+++ src/window_main/components/settings/SectionBehaviour.tsx
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/camelcase */
 import React from "react";
-import Toggle from "../Toggle";
-import Input from "../Input";
+import Toggle from "../misc/Toggle";
+import Input from "../misc/Input";
 import { ipcSend } from "../../renderer-util";
 import { useSelector } from "react-redux";
 import { AppState } from "../../../shared/redux/appState";
