+++ src/window_main/components/settings/sectionVisual.tsx
@@ -7,9 +7,9 @@ import ReactSelect from "../../../shared/ReactSelect";
 import { CARD_TILE_ARENA, CARD_TILE_FLAT } from "../../../shared/constants";
 import CardTile from "../../../shared/CardTile";
 import db from "../../../shared/database";
-import Input from "../Input";
+import Input from "../misc/Input";
 import useColorPicker from "../../hooks/useColorPicker";
-import Slider from "../Slider";
+import Slider from "../misc/Slider";
 import { getCardImage } from "../../../shared/util";
 import { AppState } from "../../../shared/redux/appState";
 import { useSelector } from "react-redux";
